+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fix issue with list bullet in combination with multi column layouts
 - Allow for special characters in URI hash in Scroll Component
 - Height Match also sets a `height` in IE
+- Slideshow sets `min-height` instead of `height`
 
 ## 3.0.0 rc 16 (September 12, 2018)